WIP: move to lxml, make utils a package
All tests pass, but more analysis is needed.
